refactoring all delete methods

+++ src/main/java/com/library/app/Service/Books/BookService.java
@@ -35,6 +35,7 @@ public class BookService {
     }
 
     public String deleteBook(BigInteger id) {
+        if (booksRepo.findById(id).isEmpty()) return null;
         booksRepo.deleteById(id);
         return "ok";
     }
